// front_end/src/app/types/Product.ts
import { Bid } from './Bid';

export type Product = {
  _id: string;
  title: string;
  price: number;
  description: string;
  category: {
    _id: string;
    name: string;
  }
  image: string;
  isShow: boolean;
  bids: Bid[];
  startAt: Date;
  endAt: Date;
  bidPriceMax: number;
};

export type ProductForm = {
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
  isShow: boolean;
};